django/requirements.txt
gunicorn==23.0.0
Django==5.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-cors-headers==4.5.0
Markdown==3.7
django-filter==24.3
djangorestframework-simplejwt==5.3.1
dj-database-url==2.3.0
drf-spectacular==0.27.2